Rename java packages from screenhandler to menu
