--- fabric-menu-api-v1/src/testmod/java/net/fabricmc/fabric/test/screenhandler/block/BoxBlock.java
@@ -1,86 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.test.screenhandler.block;
-
-import com.mojang.serialization.MapCodec;
-
-import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
-import net.minecraft.server.level.ServerLevel;
-import net.minecraft.world.Container;
-import net.minecraft.world.Containers;
-import net.minecraft.world.InteractionResult;
-import net.minecraft.world.MenuProvider;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.BaseEntityBlock;
-import net.minecraft.world.level.block.RenderShape;
-import net.minecraft.world.level.block.entity.BlockEntity;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.phys.BlockHitResult;
-
-public class BoxBlock extends BaseEntityBlock {
-	public BoxBlock(Properties settings) {
-		super(settings);
-	}
-
-	@Override
-	protected MapCodec<? extends BaseEntityBlock> codec() {
-		throw new UnsupportedOperationException("not implemented yet");
-	}
-
-	@Override
-	public InteractionResult useWithoutItem(BlockState state, Level world, BlockPos pos, Player player, BlockHitResult hit) {
-		if (!world.isClientSide()) {
-			MenuProvider screenHandlerFactory = state.getMenuProvider(world, pos);
-
-			if (screenHandlerFactory != null) {
-				player.openMenu(screenHandlerFactory);
-			}
-		}
-
-		return InteractionResult.SUCCESS;
-	}
-
-	@Override
-	public BlockEntity newBlockEntity(BlockPos blockPos, BlockState blockState) {
-		return new BoxBlockEntity(blockPos, blockState);
-	}
-
-	@Override
-	public RenderShape getRenderShape(BlockState state) {
-		return RenderShape.MODEL;
-	}
-
-	@Override
-	public void affectNeighborsAfterRemoval(BlockState state, ServerLevel world, BlockPos pos, boolean moved) {
-		BlockEntity be = world.getBlockEntity(pos);
-
-		if (be instanceof Container) {
-			Containers.dropContents(world, pos, (Container) be);
-			world.updateNeighbourForOutputSignal(pos, this);
-		}
-
-		super.affectNeighborsAfterRemoval(state, world, pos, moved);
-	}
-
-	@Override
-	public int getAnalogOutputSignal(BlockState state, Level world, BlockPos pos, Direction direction) {
-		return AbstractContainerMenu.getRedstoneSignalFromBlockEntity(world.getBlockEntity(pos));
-	}
-}
