--- fabric-menu-api-v1/src/testmod/java/net/fabricmc/fabric/test/screenhandler/block/BoxBlockEntity.java
@@ -1,69 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.test.screenhandler.block;
-
-import net.minecraft.core.BlockPos;
-import net.minecraft.core.NonNullList;
-import net.minecraft.network.chat.Component;
-import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.block.entity.RandomizableContainerBlockEntity;
-import net.minecraft.world.level.block.state.BlockState;
-
-import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerFactory;
-import net.fabricmc.fabric.test.screenhandler.ScreenHandlerTest;
-import net.fabricmc.fabric.test.screenhandler.screen.BoxScreenHandler;
-
-public class BoxBlockEntity extends RandomizableContainerBlockEntity implements ExtendedScreenHandlerFactory<BlockPos> {
-	private NonNullList<ItemStack> items = NonNullList.withSize(getContainerSize(), ItemStack.EMPTY);
-
-	public BoxBlockEntity(BlockPos blockPos, BlockState blockState) {
-		super(ScreenHandlerTest.BOX_ENTITY, blockPos, blockState);
-	}
-
-	@Override
-	protected NonNullList<ItemStack> getItems() {
-		return items;
-	}
-
-	@Override
-	protected void setItems(NonNullList<ItemStack> list) {
-		this.items = list;
-	}
-
-	@Override
-	protected Component getDefaultName() {
-		return Component.translatable(getBlockState().getBlock().getDescriptionId());
-	}
-
-	@Override
-	protected AbstractContainerMenu createMenu(int syncId, Inventory playerInventory) {
-		return new BoxScreenHandler(syncId, playerInventory, this);
-	}
-
-	@Override
-	public int getContainerSize() {
-		return 3 * 3;
-	}
-
-	@Override
-	public BlockPos getScreenOpeningData(ServerPlayer player) {
-		return worldPosition;
-	}
-}
