--- fabric-menu-api-v1/src/testmod/java/net/fabricmc/fabric/test/screenhandler/item/BagItem.java
@@ -1,47 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.test.screenhandler.item;
-
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResult;
-import net.minecraft.world.MenuProvider;
-import net.minecraft.world.SimpleMenuProvider;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.Level;
-
-import net.fabricmc.fabric.test.screenhandler.screen.BagScreenHandler;
-
-public class BagItem extends Item {
-	public BagItem(Properties settings) {
-		super(settings);
-	}
-
-	@Override
-	public InteractionResult use(Level world, Player user, InteractionHand hand) {
-		ItemStack stack = user.getItemInHand(hand);
-		user.openMenu(createScreenHandlerFactory(stack));
-		return InteractionResult.SUCCESS;
-	}
-
-	private MenuProvider createScreenHandlerFactory(ItemStack stack) {
-		return new SimpleMenuProvider((syncId, inventory, player) -> {
-			return new BagScreenHandler(syncId, inventory, new BagInventory(stack));
-		}, stack.getHoverName());
-	}
-}
